Rename classes to make it more obvious
