+++ Assets/Editor/Editor/CustomToolbarRight.cs
@@ -7,17 +7,15 @@ namespace UnityToolbarExtender
 {
     
     [InitializeOnLoad]
-    public static class RestartPlayMode
+    public static class CustomToolbarRight
     {
-
-        static RestartPlayMode()
+        static CustomToolbarRight()
         {
             ToolbarExtender.RightToolbarGUI.Add(OnToolbarGUI);
         }
+
         static void OnToolbarGUI()
         {
-
-            
             EditorGUIUtility.SetIconSize(new Vector2(17,17));
             if (GUILayout.Button(new GUIContent((Texture2D)AssetDatabase.LoadAssetAtPath("Assets/Editor/Icons/LookDevResetEnv@2x.png", typeof(Texture2D))), ToolbarStyles.commandButtonStyle))
             {
@@ -26,10 +24,8 @@ namespace UnityToolbarExtender
                     SceneManager.LoadScene(SceneManager.GetActiveScene().name);
                 }
             }
-
             
             Time.timeScale = EditorGUILayout.Slider("", Time.timeScale, 1, 20,GUILayout.Width(150));
-
         }
     }
 }
