+++ Assets/Editor/Editor/ToolbarStyles.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace UnityToolbarExtender
+{
+    static class ToolbarStyles
+    {
+        public static readonly GUIStyle commandButtonStyle;
+
+        static ToolbarStyles()
+        {
+            commandButtonStyle = new GUIStyle("Command")
+            {
+                fontSize = 16,
+                alignment = TextAnchor.MiddleCenter,
+                imagePosition = ImagePosition.ImageAbove,
+                fontStyle = FontStyle.Bold
+            };
+        }
+    }
+}
