+++ Assets/Editor/Editor/ToolbarStyles.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47ad6f372df56df46883727f1cefad85
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
